fix: release 1.4.0 — coach markers, variation selection & book improvements
Core fixes:
- Fixed coach tools markers misalignment with board squares after
  hiding coordinates and when flipping the board
- Fixed variation selection at a main line fork: the chosen variation
  is now always applied instead of falling back to the main line
- Fixed navigation panel buttons in book mode
- Fixed Polyglot book loading (negative number during filtering now
  handled correctly)
- Fixed GitHub page opening on Linux (Debian)
- Fixed move display when dragging a piece during new variation creation
- Fixed indexing when saving a position to a non-indexed PGN file

Polyglot book improvements:
- Added book editing: new variations can be added during the game
- Added undo of an added move before saving the book
- Added Save Book button
- Redesigned context menu for correct deletion of variations, moves
  and tree rebuilding
- Accelerated book loading via MappedByteBuffer — large and small
  books now load almost instantly, even on HDD
- Book mode: engine analysis is allowed, move execution is disabled

Other:
- Added board flip during position setup
- Removed excessive logging and debug messages
- Updated CHANGELOG for 1.4.0 (Russian, English, Chinese)
- Bumped Actions to Node.js 24 compatible versions
  (checkout@v5, setup-java@v5, upload-artifact@v5)

+++ src/main/java/Khrypach/Andrey/chess/kletka/gui/book/PolyglotEntry.java
@@ -35,7 +35,7 @@ import com.github.bhlangonijr.chesslib.Square;
  * @param learn  Информация для обучения
  * @param games  Количество партий
  */
-public record PolyglotEntry(long key, short move, short weight, int learn, int games) {
+public record PolyglotEntry(long key, short move, int weight, int learn, int games) {
 
     /**
      * Получает начальную клетку из 16-битного хода
